pyproject.toml
[project]
name = "robotpuzzle"
version = "0.1.0"
description = "Puzzle solving robot application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python",
    "numpy",
    "customtkinter",
    "Pillow",
    "scipy",
    "pyserial>=3.5",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
